requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<3.16
drf-spectacular>=0.28,<0.29
psycopg[binary]>=3.2,<3.3
redis>=5.2,<5.3
celery[redis]>=5.4,<5.5
django-environ>=0.11,<0.12
openpyxl>=3.1,<3.2
gunicorn>=23.0,<24.0
pytest>=8.3,<8.4
pytest-django>=4.9,<4.10
factory-boy>=3.3,<3.4
